// types/content.ts
export type NavItem = string;

export type Service = {
  title: string;
  description: string;
};

export type TeamMember = {
  name: string;
  role: string;
  img: string;
};

export type Benefit = {
  title: string;
  description: string;
};

export type Testimonial = {
  text: string;
  author: string;
};

export type Blog = {
  title: string;
  date: string;
  type: string;
  excerpt: string;
};